+++ src/mutex/PHPRedisMutex.php
@@ -29,12 +29,12 @@ class PHPRedisMutex extends RedisMutex
      * called already.
      *
      * @param array<\Redis|\RedisCluster> $redisAPIs The Redis connections.
-     * @param string $name The lock name.
-     * @param int $timeout The time in seconds a lock expires after. Default is
-     *                     3 seconds.
+     * @param string                      $name      The lock name.
+     * @param float                       $timeout   The time in seconds a lock expires after. Default is
+     *                                               3 seconds.
      * @throws \LengthException The timeout must be greater than 0.
      */
-    public function __construct(array $redisAPIs, string $name, int $timeout = 3)
+    public function __construct(array $redisAPIs, string $name, float $timeout = 3)
     {
         parent::__construct($redisAPIs, $name, $timeout);
     }
@@ -43,7 +43,7 @@ class PHPRedisMutex extends RedisMutex
      * @param \Redis|\RedisCluster $redisAPI The Redis or RedisCluster connection.
      * @throws LockAcquireException
      */
-    protected function add($redisAPI, string $key, string $value, int $expire): bool
+    protected function add($redisAPI, string $key, string $value, float $expire): bool
     {
         /** @var \Redis $redisAPI */
         try {
